Add files via upload
An example output file with formatted cross section data. The help file includes the grep command use for a generic SMR output file. This will need to be modified for the reflector case. 

I will begin working on a bash script to do this automatically each time casmo is ran for this case.